kpa_flask_sqlite_assignment/requirements.txt
Flask==2.3.2
Flask_SQLAlchemy==3.1.1
python-dotenv==1.0.1
